+++ templates/3dsecure/form.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  * @package  Mastercard
- * @version  GIT: @1.4.6@
+ * @version  GIT: @1.4.7@
  * @link     https://github.com/fingent-corp/gateway-woocommerce-mastercard-module/
  */
 
@@ -29,7 +29,7 @@
  */
 ?>
 <!doctype html>
-<html>
+<html <?php language_attributes(); ?>>
 	<head>
 		<title><?php esc_html_e( 'Processing Secure Payment', 'mastercard' ); ?></title>
 		<meta http-equiv="content-type" content="text/html; charset=utf-8"/>